Add files via upload
Pain.... is the best word to explain this project for my own website. it was a good challenge and it was a lot harder than I thought. I did a lot of research  about this topic. The has been some bugs that I have tried the life of me to fix but I don't know how to fix them.